+++ Entities/MovableObject.cpp
@@ -109,8 +109,8 @@ void MovableObject::Clear()
 
 	m_ProvidesPieMenuContext = false;
 
-    m_UpdateFrames = 0;
-    m_UpdatePeriod = 1;
+	m_SimUpdatesBetweenScriptedUpdates = 1;
+    m_SimUpdatesSinceLastScriptedUpdate = 0;
 }
 
 
@@ -136,8 +136,6 @@ int MovableObject::Create()
 	m_MOIDHit = g_NoMOID;
 	m_TerrainMatHit = g_MaterialAir;
 	m_ParticleUniqueIDHit = 0;
-
-    m_UpdateFrames = 0;
 
     g_MovableMan.RegisterObject(this);
 
@@ -174,8 +172,6 @@ int MovableObject::Create(const float mass,
 	m_MOIDHit = g_NoMOID;
 	m_TerrainMatHit = g_MaterialAir;
 	m_ParticleUniqueIDHit = 0;
-
-    m_UpdateFrames = 0;
 
 	g_MovableMan.RegisterObject(this);
 
@@ -260,7 +256,8 @@ int MovableObject::Create(const MovableObject &reference)
 	m_TerrainMatHit = reference.m_TerrainMatHit;
 	m_ParticleUniqueIDHit = reference.m_ParticleUniqueIDHit;
 
-    m_UpdatePeriod = reference.m_UpdatePeriod;
+	m_SimUpdatesBetweenScriptedUpdates = reference.m_SimUpdatesBetweenScriptedUpdates;
+	m_SimUpdatesSinceLastScriptedUpdate = reference.m_SimUpdatesSinceLastScriptedUpdate;
 
 	m_UniqueID = MovableObject::GetNextUniqueID();
 	g_MovableMan.RegisterObject(this);
@@ -410,8 +407,8 @@ int MovableObject::ReadProperty(const std::string_view &propName, Reader &reader
         reader >> m_ApplyWoundBurstDamageOnCollision;
 	else if (propName == "IgnoreTerrain")
 		reader >> m_IgnoreTerrain;
-    else if (propName == "UpdatePeriod")
-        reader >> m_UpdatePeriod;
+    else if (propName == "SimUpdatesBetweenScriptedUpdates")
+        reader >> m_SimUpdatesBetweenScriptedUpdates;
 	else
         return SceneObject::ReadProperty(propName, reader);
 
@@ -1018,17 +1015,17 @@ void MovableObject::Update()
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 int MovableObject::UpdateScripts() {
-    m_UpdateFrames++;
+    m_SimUpdatesSinceLastScriptedUpdate++;
     
     if (m_AllLoadedScripts.empty() || m_ScriptPresetName.empty()) {
         return -1;
     }
 
-    if (m_UpdateFrames < m_UpdatePeriod) {
+    if (m_SimUpdatesSinceLastScriptedUpdate < m_SimUpdatesBetweenScriptedUpdates) {
         return 1;
     }
 
-    m_UpdateFrames = 0;
+	m_SimUpdatesSinceLastScriptedUpdate = 0;
 
 
     int status = !g_LuaMan.ExpressionIsTrue(m_ScriptPresetName, false) ? ReloadScripts() : 0;